api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "asyncpg>=0.31.0",
    "authlib>=1.6.5",
    "databases>=0.9.0",
    "fastapi>=0.123.3",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
    "PyJWT>=2.8.0",
    "itsdangerous>=2.2.0",
    "openai-agents>=0.6.1",
    "qdrant-client>=1.16.1",
]
